feat: retire integrations utils helper functions
